+++ app/middlewares/validator/authControllerValidators.js
@@ -1,9 +1,10 @@
 'use strict';
 
 const { userValidators } = require('./definition/allValidators');
+const validationResultCheck = require('./definition/validationResultCheck');
 
 const signupValidators = [userValidators.email, userValidators.password('password')];
 
 module.exports = {
-    signupValidators
+    signupValidators: signupValidators.concat(validationResultCheck)
 };
