+++ app/middlewares/validator/scheduleControllerValidators.js
@@ -1,11 +1,12 @@
 'use strict';
 
 const { scheduleValidators } = require('./definition/allValidators');
+const validationResultCheck = require('./definition/validationResultCheck');
 
 const readScheduleValidators = [scheduleValidators.date];
 const updateScheduleValidators = [scheduleValidators.date, scheduleValidators.startTime, scheduleValidators.endTime];
 
 module.exports = {
-    readScheduleValidators,
-    updateScheduleValidators
+    readScheduleValidators: readScheduleValidators.concat(validationResultCheck),
+    updateScheduleValidators: updateScheduleValidators.concat(validationResultCheck)
 };
